chore: build watch
